refactor: keep the productions with the minimizers that read them
The wrap and the read-back move into `utils.js` beside the four minimizers, and
the three productions JavaScript is handed out in are named there: a classic
script, a module, and the function body an event handler holds. A minify
function reaches a worker as its source, so each still requires them rather
than reading this module's scope.
